fix: make `reader-dyn--{next,prev}-page' return `t' on success.
* In: a8b0366b1b925f569d9096606df304a2db6e8209, we had stopped forcing
mode-line updates on every buffer when the page has changed, this
caused the unintentional side effect of the functions always returning
nil, even when the command succeeds. This was subtly breaking all
other commands which relied on the return value of the functions.

* Replace: `if-let*' with `when' since the "status" returned by
`reader-dyn--{next,prev}-page' is always going to be `t', or `nil'.
There is not point in catching it.

* Should close #157, as well. Since that PR is also trying to solve
this problem. In fact, I noticed this bug due to that PR.